Updates to existing scripts
Several updates to the scripts that generate figure 3 and 5 intermediaries. Several of the files have multiple functions to generate different versions of the plots. The OSF scripts are currently set to produce the wrong version of the plot, several other updates of that kind were made to aid in regenerating the figures shown in the paper.