Anonymized data: code update
Worked on anonymizing our data generation process to protect the user ids of individuals on twitter that were not affiliated with large media companies or were not political public figures.

The code updates allow, in many cases, for an anonymization flag that converts users into anonymous identifiers. The rest of the code can consume these as they did originally, with minor modifications here and there to account for some figures hard coding some stuff.